refactor(utils): move AST helpers into dedicated folder
