+++ .changeset/move-ast-helpers.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': patch
+---
+
+move AST helpers into src/utils/ast directory
