+++ src/rules/deprecation.ts
@@ -1,7 +1,7 @@
 import ts from 'typescript';
 import { z } from 'zod';
 import type { RuleModule } from '../core/types.js';
-import { isInNonStyleJsx } from '../utils/is-in-non-style-jsx.js';
+import { isInNonStyleJsx } from '../utils/ast/index.js';
 
 export const deprecationRule: RuleModule = {
   name: 'design-system/deprecation',
