+++ src/utils/ast/is-style-name.ts
